# Query result cache to speed up query processing for rarely mutable datasets.